backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0 
python-multipart==0.0.22 
requests==2.32.5
gradio_client==2.0.3
python-dotenv==1.2.1